moved data-related functions around
