+++ kalman_watch/kalman_perigee_mon.py
@@ -2,14 +2,9 @@
 """Watch Kalman star data during perigee passages."""
 
 import argparse
-import calendar
-import functools
-import json
 import os
-import re
 from itertools import cycle
 from pathlib import Path
-from typing import List, Union
 
 import astropy.units as u
 import numpy as np
@@ -17,18 +12,24 @@ import plotly.express as px
 import plotly.graph_objects as pgo
 from acdc.common import send_mail
 from astropy.table import Table, vstack
-from cheta.fetch import MSIDset
-from cheta.utils import logical_intervals
 from cxotime import CxoTime
 from jinja2 import Template
-from kadi.commands import get_observations
-from kadi.commands.commands_v2 import get_cmds
-from kadi.commands.states import get_states, reduce_states
 from plotly.subplots import make_subplots
 from ska_helpers.logging import basic_logger
 from ska_helpers.run_info import log_run_info
 
 from kalman_watch import __version__
+from kalman_watch.kalman_watch_data import (
+    PERIGEES_INDEX_TABLE_PATH,
+    EVT_PERIGEE_DATA_PATH,
+    EVT_PERIGEE_DIR_PATH,
+    PERIGEES_DIR_PATH,
+    get_dirname,
+    EventPerigee,
+    get_stats,
+    get_evts_perigee,
+    read_kalman_stats,
+)
 
 LOGGER = basic_logger(__name__, level="INFO")
 
@@ -37,26 +38,6 @@ FILE_DIR = Path(__file__).parent
 
 # Sub-sample attitude error by 8 to reduce the number of points.
 ATT_ERR_SUBSAMP = 8
-
-
-def PERIGEES_DIR_PATH(data_dir: str) -> Path:
-    return Path(data_dir) / "perigees"
-
-
-def PERIGEES_INDEX_TABLE_PATH(data_dir: str) -> Path:
-    return PERIGEES_DIR_PATH(data_dir) / "kalman_perigees.ecsv"
-
-
-def EVT_PERIGEE_DATA_PATH(data_dir: str, evt: "EventPerigee") -> Path:
-    return EVT_PERIGEE_DIR_PATH(data_dir, evt) / "data.npz"
-
-
-def EVT_PERIGEE_INFO_PATH(data_dir: str, evt: "EventPerigee") -> Path:
-    return EVT_PERIGEE_DIR_PATH(data_dir, evt) / "info.json"
-
-
-def EVT_PERIGEE_DIR_PATH(data_dir: str, evt: "EventPerigee"):
-    return PERIGEES_DIR_PATH(data_dir) / evt.dirname
 
 
 def INDEX_DETAIL_PATH():
@@ -89,16 +70,6 @@ COLOR_CYCLE = [
     "#bcbd22",  # curry yellow-green
     "#17becf",  # blue-teal
 ]
-
-
-def get_dirname(date: Union[CxoTime, None]) -> str:
-    if date is None:
-        out = ""
-    else:
-        ymdhms = date.ymdhms
-        mon = calendar.month_abbr[ymdhms.month]
-        out = f"{ymdhms['year']}/{mon}-{ymdhms['day']:02d}"
-    return out
 
 
 def get_opt() -> argparse.ArgumentParser:
@@ -170,131 +141,14 @@ def main(sys_args=None):
         ):
             if stat["perigee"] < start.date:
                 break
-            evt_tmp = EventPerigee(stat["rad_entry"], stat["perigee"], stat["rad_exit"])
+            evt_tmp = EventPerigeeMon(stat["rad_entry"], stat["perigee"], stat["rad_exit"])
             path = EVT_PERIGEE_DATA_PATH(opt.data_dir, evt_tmp)
-            evt = EventPerigee.from_npz(path)
+            evt = EventPerigeeMon.from_npz(path)
             evt.prev_date = date_prev
             evt.next_date = date_next
             evt.make_detail_page(opt)
 
         make_index_list_pages(opt, stats_all)
-
-
-def read_kalman_stats(opt, from_info=False) -> Table:
-    """Read kalman stats from file or from event info.json files.
-
-    If ``from_info`` is True, read all individual event info files instead of
-    the data file. This is also tried if the kalman stats file does not exist,
-    which allows re-generating the kalman stats file.
-    """
-    path = PERIGEES_INDEX_TABLE_PATH(opt.data_dir)
-    if path.exists() and not from_info:
-        LOGGER.info(f"Reading kalman perigee data from {path}")
-        kalman_stats = Table.read(path)
-    else:
-        rows = []
-        # Look for files like 2019/Jan-12/info.json
-        for info_file in PERIGEES_DIR_PATH(opt.data_dir).glob("????/??????/info.json"):
-            if re.search(r"\d{4}/\w{3}-\d{2}/info\.json", info_file.as_posix()):
-                LOGGER.info(f"Reading kalman perigee data from {info_file}")
-                info = json.loads(info_file.read_text())
-                rows.append(info)
-
-        LOGGER.info(f"No kalman perigee stats data found at {path}")
-        LOGGER.info(f"Creating new table from {len(rows)} info files")
-        kalman_stats = Table(rows=rows)
-        if rows:
-            kalman_stats.sort("perigee", reverse=True)
-
-    return kalman_stats
-
-
-def get_evts_perigee(
-    start: CxoTime, stop: CxoTime, stats_prev: Table
-) -> List["EventPerigee"]:
-    """
-    Get the perigee events within start/stop.
-
-    This selects perigees within start/stop and then finds the span of
-    ERs (obsid > 38000) within +/- 12 hours of perigee.
-
-    :param start: CxoTime
-        Start of date range
-    :param stop: CxoTime
-        End of date range
-    :param stats_prev: Table
-        Previous kalman stats table
-    :returns: list of PerigeeEvent
-        List of PerigeeEvent objects
-    """
-    LOGGER.info(f"Getting perigee events between {start} and {stop}")
-    # event_types = ["EEF1000", "EPERIGEE", "XEF1000"]
-    cmds_perigee = get_cmds(
-        start=start, stop=stop, type="ORBPOINT", event_type="EPERIGEE"
-    )
-
-    # Find contiguous intervals of ERs (obsid > 38000)
-    states = get_states(start - 3 * u.day, stop + 3 * u.day, state_keys=["obsid"])
-    states["obsid"] = np.where(states["obsid"] > 38000, 1, 0)
-    states = reduce_states(states, state_keys=["obsid"], merge_identical=True)
-
-    dirnames_prev = stats_prev["dirname"] if len(stats_prev) > 0 else []
-
-    events = []
-    for cmd in cmds_perigee:
-        t_perigee = cmd["time"]
-        ok = (states["tstart"] <= t_perigee) & (t_perigee < states["tstop"])
-        n_ok = np.count_nonzero(ok)
-        if n_ok == 0:
-            LOGGER.warning(
-                "WARNING: No ER observations found covering perigee at"
-                f" {CxoTime(t_perigee).date}"
-            )
-            continue
-        elif n_ok > 1:
-            raise ValueError(
-                "Found multiple states covering perigee at"
-                f" {CxoTime(t_perigee).date} (this really should not happen, this must"
-                " be a bug"
-            )
-
-        t_rad_entry = max(states["tstart"][ok][0], t_perigee - 20000)
-        t_rad_exit = min(states["tstop"][ok][0], t_perigee + 20000)
-
-        event = EventPerigee(
-            rad_entry=t_rad_entry,
-            perigee=cmd["date"],
-            rad_exit=t_rad_exit,
-        )
-
-        if event.dirname in dirnames_prev:
-            # If the event is already in the previous kalman stats table then
-            # move on silently.
-            continue
-
-        if event.tlm is not None:
-            events.append(event)
-        else:
-            LOGGER.info(f"No TLM found for perigee event at {event.perigee}, skipping")
-            continue
-
-    LOGGER.info(f"Found {len(events)} new perigee event(s)")
-    return events
-
-
-def get_stats(evts_perigee) -> Table:
-    """
-    Get the kalman perigee stats for the given events.
-
-    :param evts_perigee: list of PerigeeEvent
-        List of PerigeeEvent objects
-    :returns: Table
-        Table of kalman perigee stats
-    """
-    rows = [evt.info for evt in reversed(evts_perigee)]
-
-    out = Table(rows=rows)
-    return out
 
 
 def make_index_list_pages(opt, stats_all: Table) -> None:
@@ -395,288 +249,12 @@ def send_process_mail(opt, evts_perigee):
     send_mail(LOGGER, opt, subject, text, __file__)
 
 
-@functools.lru_cache()
-def get_rad_table():
-    rad_table = Table.read(Path(os.environ["SKA"]) / "data" / "stk_radiation" / "rad_data_2022:003:12:00:00.000-2025:365:11:59:59.000.fits")
-    rad_table["time"] = CxoTime(rad_table["time"])
-    return rad_table
-
-
-class EventPerigee:
+class EventPerigeeMon(EventPerigee):
     """Class for tracking Kalman star data through perigee."""
 
     def __init__(self, rad_entry, perigee, rad_exit):
-        self.rad_entry = CxoTime(rad_entry)
-        self.perigee = CxoTime(perigee)
-        self.rad_exit = CxoTime(rad_exit)
-        self.prev_date = None
-        self.next_date = None
-
-    @property
-    def prev_date(self):
-        return self._prev_date
-
-    @prev_date.setter
-    def prev_date(self, value):
-        self._prev_date = value if value is None else CxoTime(value)
-
-    @property
-    def next_date(self):
-        return self._next_date
-
-    @next_date.setter
-    def next_date(self, value):
-        self._next_date = value if value is None else CxoTime(value)
-
-    @classmethod
-    def from_npz(cls, path):
-        LOGGER.info(f"Loading perigee event from {path}")
-        with np.load(path) as npz_data:
-            data = dict(npz_data)
-        obj = cls(data["rad_entry"], data["perigee"], data["rad_exit"])
-
-        bad = data["aokalstr"] == -1
-        data["aokalstr"] = data["aokalstr"].astype(np.float64)
-        data["aokalstr"][bad] = np.nan
-
-        for slot in range(8):
-            data[f"aca_track{slot}"] = (data["aca_track"] & (1 << slot)).astype(bool)
-            data[f"aca_ir{slot}"] = (data["aca_ir"] & (1 << slot)).astype(bool)
-
-        data["times"] = obj.perigee.secs + data["perigee_times"].astype(np.float64)
-
-        obj._data = data
-        obj._obss = Table(data["obss"])
-        return obj
-
-    def __repr__(self):
-        return (
-            f"EventPerigee(rad_entry={self.rad_entry.date!r},"
-            f" perigee={self.perigee.date!r},"
-            f" rad_exit={self.rad_exit.date!r})"
-        )
-
-    @property
-    def dirname(self):
-        return get_dirname(self.perigee)
-
-    @property
-    def obss(self):
-        if not hasattr(self, "_obss"):
-            LOGGER.info(f"Getting observations from kadi commands for {self.dirname}")
-            obss = get_observations(start=self.rad_entry, stop=self.rad_exit)
-            self._obss = Table(obss)
-        return self._obss
-
-    @property
-    def tlm(self):
-        if not hasattr(self, "_tlm"):
-            self._tlm = self._get_tlm()
-        return self._tlm
-
-    def _get_tlm(self) -> Union[MSIDset, None]:
-        """Get telemetry for processing perigee for Kalman
-
-        :returns: MSIDset or None
-        """
-        msids = [
-            "aokalstr",
-            "aopcadmd",
-            "aoacaseq",
-            "aoaciir*",
-            "aoacrpt",
-            "aoacfct*",
-            "aoatter*",
-        ]
-        LOGGER.info(f"Getting telemetry for {self.perigee}")
-        tlm = MSIDset(msids, self.rad_entry, self.rad_exit)
-        if (
-            len(tlm["aokalstr"]) == 0
-            or self.rad_exit.cxcsec - tlm["aokalstr"].times[-1] > 100
-        ):
-            return None
-
-        tlm["aokalstr"].vals = tlm["aokalstr"].vals.astype(np.float64)
-
-        # Reduce everything to the first ACA values during NPNT/KALM
-        ok = tlm["aoacrpt"].vals.astype(int) == 0
-        tlm.interpolate(
-            times=tlm["aokalstr"].times[ok], bad_union=False, filter_bad=False
-        )
-
-        bad = (
-            (tlm["aopcadmd"].vals != "NPNT")
-            | (tlm["aoacaseq"].vals != "KALM")
-            | tlm["aokalstr"].bads
-        )
-        tlm["aokalstr"].vals[bad] = np.nan
-        for axis in range(1, 4):
-            tlm[f"aoatter{axis}"].vals[bad] = np.nan
-
-        tlm.perigee_times = tlm.times - self.perigee.cxcsec
-
-        return tlm
-
-    @property
-    def data(self):
-        """Get data for processing perigee"""
-        if not hasattr(self, "_data"):
-            self._data = self._get_data()
-        return self._data
-
-    def _get_data(self) -> dict:
-        """Get data for processing perigee for Kalman
-
-        :returns: dict of data
-        """
-        LOGGER.debug(f"Setting data property for {self.dirname}")
-        data = {}
-        for axis in range(1, 4):
-            # Subsample by 8 since this does not vary quickly
-            data[f"aoatter{axis}"] = (
-                self.tlm[f"aoatter{axis}"].vals[::ATT_ERR_SUBSAMP].astype(np.float32)
-            )
-        data["aokalstr"] = self.tlm["aokalstr"].vals
-        # fmt: off
-        data["npnt_kalm"] = (
-            (self.tlm["aopcadmd"].vals == "NPNT")
-            & (self.tlm["aoacaseq"].vals == "KALM")
-        )
-        # fmt: on
-        for slot in range(8):
-            data[f"aca_track{slot}"] = self.tlm[f"aoacfct{slot}"].vals == "TRAK"
-            data[f"aca_ir{slot}"] = self.tlm[f"aoaciir{slot}"].vals == "ERR"
-        data["times"] = self.tlm["aokalstr"].times
-        data["perigee_times"] = self.tlm.perigee_times.astype(np.float32)
-        data["perigee"] = self.perigee.date
-        data["rad_entry"] = self.rad_entry.date
-        data["rad_exit"] = self.rad_exit.date
-        data["obss"] = self.obss.as_array()
-
-        return data
-
-    @property
-    def info(self):
-        if not hasattr(self, "_info"):
-            self._info = self._get_info()
-        return self._info
-
-    def _get_info(self):
-        low_kals = self.low_kalmans
-        info = {
-            "dirname": self.dirname,
-            "perigee": self.perigee.date[:19],
-        }
-        for key in ["rad_entry", "perigee", "rad_exit"]:
-            info[key] = getattr(self, key).date
-        for nle in (3, 2, 1):
-            info[f"n{nle}_ints"] = np.count_nonzero((low_kals["n_kalstr"] == nle))
-            info[f"n{nle}_cnt"] = low_kals.meta[f"n{nle}_cnt"]
-        return info
-
-    @property
-    def predicted_kalman_drops(self):
-        if not hasattr(self, "_info"):
-            self._predicted_kalman_drops = self._get_predicted_kalman_drops()
-        return self._predicted_kalman_drops
-
-    def _get_predicted_kalman_drops(self):
-        # result of a linear fit of kalman_drops vs proton_26_300_MeV
-        b, a = 1.72306678e-09, 0.011492516864638577
-        rad_table = get_rad_table()
-        sel = (rad_table["time"] >= self.rad_entry) & (rad_table["time"] <= self.rad_exit)
-        predicted_kalman_drops = {
-            "times": (rad_table["time"][sel] - self.perigee).sec,
-            "values": a + rad_table["proton_26_300_MeV"][sel] * b,
-        }
-        return predicted_kalman_drops
-
-    def write_info(self, opt):
-        """Write info to file"""
-        path = EVT_PERIGEE_INFO_PATH(opt.data_dir, self)
-        LOGGER.info(f"Writing info to {path}")
-        path.write_text(json.dumps(self.info, indent=4))
-
-    def write_data(self, opt):
-        # Compressed version of data
-        dc = {}
-
-        # Store aokalstr as int8 with the nan's as -1
-        vals = self.data["aokalstr"].copy()
-        vals[np.isnan(vals)] = -1
-        dc["aokalstr"] = vals.astype(np.int8)
-
-        for key in (
-            "perigee_times",
-            "npnt_kalm",
-            "perigee",
-            "rad_entry",
-            "rad_exit",
-            "obss",
-            "aoatter1",
-            "aoatter2",
-            "aoatter3",
-        ):
-            dc[key] = self.data[key]
-
-        dc["aca_track"] = np.zeros(len(self.data["times"]), dtype=np.uint8)
-        dc["aca_ir"] = np.zeros(len(self.data["times"]), dtype=np.uint8)
-        for slot in range(8):
-            dc["aca_track"] |= self.data[f"aca_track{slot}"].astype(np.uint8) << slot
-            dc["aca_ir"] |= self.data[f"aca_ir{slot}"].astype(np.uint8) << slot
-
-        path = EVT_PERIGEE_DATA_PATH(opt.data_dir, self)
-        path.parent.mkdir(parents=True, exist_ok=True)
-        LOGGER.info(f"Writing perigee data to {path}")
-        np.savez_compressed(path, **dc)
-
-    @property
-    def low_kalmans(self):
-        if not hasattr(self, "_low_kalmans"):
-            self._low_kalmans = self._get_low_kalmans()
-        return self._low_kalmans
-
-    def _get_low_kalmans(self) -> Table:
-        rows = []
-        for n_kalstr, dur_limit in KALMAN_LIMITS:
-            vals = self.data["aokalstr"].copy()
-            vals[np.isnan(vals)] = 10
-            ints_low = logical_intervals(self.data["times"], vals <= n_kalstr)
-            ints_low = ints_low[ints_low["duration"] > dur_limit]
-
-            for int_low in ints_low:
-                p0 = int_low["tstart"] - self.perigee.cxcsec
-                p1 = int_low["tstop"] - self.perigee.cxcsec
-                row = dict(int_low)
-                del row["tstart"]
-                del row["tstop"]
-                row["n_kalstr"] = n_kalstr
-                row["tstart_rel"] = p0
-                row["tstop_rel"] = p1
-                rows.append(row)
-
-        if len(rows) > 0:
-            low_kalmans = Table(rows=rows)
-        else:
-            low_kalmans = Table(
-                names=[
-                    "datestart",
-                    "datestop",
-                    "duration",
-                    "n_kalstr",
-                    "tstart_rel",
-                    "tstop_rel",
-                ],
-                dtype=[str, str, float, int, float, float],
-            )
-        for col in low_kalmans.itercols():
-            if col.info.dtype.kind == "f":
-                col.info.format = ".1f"
-
-        for nle in (1, 2, 3):
-            low_kalmans.meta[f"n{nle}_cnt"] = np.count_nonzero(vals <= nle)
-
-        return low_kalmans
+        super().__init__(rad_entry, perigee, rad_exit)
+        pass
 
     def get_detail_html(self) -> str:
         fig = self.get_plot_fig()
@@ -893,150 +471,6 @@ class EventPerigee:
 
         return trace
     
-    def get_kalman_drops_nman(self):
-        # having this at top-level might create a circular import
-        from kalman_watch.monitor_win_perigee import (
-            get_manvrs_perigee, get_mon_dataset, get_kalman_drops_nman, NotEnoughImagesError
-        )
-
-        # Intervals of NMAN within 100 minutes of perigee
-        manvrs_perigee = get_manvrs_perigee(self.rad_entry, self.rad_exit)
-
-        # Get list of monitor window data for each perigee maneuver
-        mons = []
-        for manvr in manvrs_perigee:
-            try:
-                mon = get_mon_dataset(
-                    manvr["datestart"],
-                    manvr["datestop"],
-                    # the following should be configured somehow
-                    ir_thresholds_start="2023:100",
-                    ir_thresholds_stop="2023:200",
-                    data_dir="/Users/javierg/SAO/git/kalman_watch/kalman_watch3_data",
-                    cache=True,
-                )
-                mons.append(mon)
-            except NotEnoughImagesError:
-                # logger.warning(err)
-                pass
-
-        if not mons:
-            return {
-                "times": np.array([]),
-                "values": np.array([]),
-            }
-
-        kalman_drops_nman_list = [get_kalman_drops_nman(mon) for mon in mons]
-
-        nman_table = vstack([
-            table_from_perigee(kalman_drops)
-            for kalman_drops in kalman_drops_nman_list
-            if len(kalman_drops.times) > 0
-        ])
-        nman_table.sort("cxcsec")
-
-        result = {
-            "times": np.asarray(nman_table["time"]),
-            "values": np.asarray(nman_table["kalman_drops"]),
-        }
-
-        return result
-
-
-    def get_kalman_drops_npnt(self):
-
-        if self.tlm is None or len(self.data["times"]) <= 200:
-            return {
-                "times": np.array([]),
-                "values": np.array([]),
-            }
-
-        times_from_perigee, n_drops = self._get_binned_drops_from_npnt()
-        result = {
-            "times": times_from_perigee,
-            "values": n_drops,
-        }
-
-        return result
-
-
-    def _get_binned_drops_from_npnt(self) -> tuple[np.ndarray, np.ndarray]:
-        """Get the fraction of IR flags per "minute" from NPNT telemetry.
-
-        Here a "minute" is really 60 * 1.025 seconds = 61.5, or 1.025 minutes. This
-        corresponds to exactly 30 ACA image readouts (2.05 sec per image) per "minute".
-
-        Parameters
-        ----------
-        ep : EventPerigee
-            Perigee event object with relevant ACA telemetry from one perigee
-
-        Returns
-        -------
-        time_means : np.ndarray
-            Array of mean time from perigee (sec) in each bin
-        ir_flag_fracs : np.ndarray
-            Array of fraction of IR flags set in each bin
-        """
-        # Select only data in AOPCADMD in NPNT and AOACASEQ in KALM
-        npnt_kalm = self.data["npnt_kalm"]
-        # Time from perigee in seconds
-        times_from_perigee = self.data["perigee_times"][npnt_kalm]
-        if len(times_from_perigee) == 0:
-            return np.array([]), np.array([])
-        ir_count = np.zeros(times_from_perigee.shape, dtype=int)
-        n_samp = np.zeros(times_from_perigee.shape, dtype=int)
-
-        # Count number of IR flags set for each slot when slot is tracking
-        for slot in range(8):
-            ir_count[:] += np.where(
-                self.data[f"aca_ir{slot}"][npnt_kalm]
-                & self.data[f"aca_track{slot}"][npnt_kalm],
-                1,
-                0,
-            )
-
-        # Count number of slot-samples when slot is tracking
-        for slot in range(8):
-            n_samp[:] += np.where(
-                self.data[f"aca_track{slot}"][npnt_kalm],
-                1,
-                0,
-            )
-
-        tbl = Table()
-        tbl["idx"] = (times_from_perigee // (60 * 1.025)).astype(int)
-        tbl["times_from_perigee"] = times_from_perigee
-        tbl["ir_count"] = ir_count
-        tbl["n_samp"] = n_samp
-        tbl_grp = tbl.group_by("idx")
-
-        time_means = []
-        ir_flag_fracs = []
-        for i0, i1 in zip(tbl_grp.groups.indices[:-1], tbl_grp.groups.indices[1:]):
-            # For a fully-sampled "minute" there are 30 ACA telemetry samples (2.05 sec per
-            # sample) times 8 slots = 240 potential samples. Require at least half of
-            # samples in a "minute" in order to return a count, otherwise just no data.
-            n_samp = np.sum(tbl_grp["n_samp"][i0:i1])
-            if n_samp > 120:
-                time_means.append(np.mean(tbl_grp["times_from_perigee"][i0:i1]))
-                # Calculate fraction of available samples that have IR flag set
-                ir_flag_fracs.append(np.sum(tbl_grp["ir_count"][i0:i1]) / n_samp)
-
-        return np.array(time_means), np.array(ir_flag_fracs)
-
-
-def table_from_perigee(perigee):
-    perigee_time = CxoTime(perigee.perigee_date).cxcsec
-    table = Table()
-    table["time"] = perigee.times
-    table["kalman_drops"] = perigee.kalman_drops
-    table["perigee"] = perigee.perigee_date
-    table["perigee_cxcsec"] = perigee_time
-    # why does the following give a different (and wrong) result?
-    # table["cxcsec"] = perigee_time + perigee.times
-    table["cxcsec"] = table["perigee_cxcsec"] + perigee.times
-    return table
 
 if __name__ == "__main__":
     import os
